requirements.txt
imagecodecs
pyyaml
scipy
termcolor
tifffile
timm
torchshow
wandb
yacs